[refactor] Move all code into src dir
