+++ lib/src/activity/trial_response/trial_response_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/mdigits_controller.dart';
+import 'package:mdigits/src/mdigits_controller.dart';
 
 /// Manage responses
 class TrialResponseController extends GetxController {
