+++ lib/src/activity/trial_stim/trial_stim_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/common/default_appbar.dart';
-import 'package:mdigits/common/centeredbox.dart';
-import 'package:mdigits/activity/trial_stim/stim_controller.dart';
+import 'package:mdigits/src/common/default_appbar.dart';
+import 'package:mdigits/src/common/centeredbox.dart';
+import 'package:mdigits/src/activity/trial_stim/stim_controller.dart';
 
 /// Screen to present stim to participants
 class TrialStimView extends StatelessWidget {
