+++ lib/src/common/default_text.dart
