+++ lib/src/end/end_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:mdigits/common/centeredbox.dart';
-import 'package:mdigits/common/default_appbar.dart';
-import 'package:mdigits/common/default_text.dart';
+import 'package:mdigits/src/common/centeredbox.dart';
+import 'package:mdigits/src/common/default_appbar.dart';
+import 'package:mdigits/src/common/default_text.dart';
 
 /// Screen for instructing the participant the task has ended.
 /// Doesn't allow going back or restarting so the experimenter has control
