+++ lib/src/errors/errors.dart
