+++ lib/src/models/trial_data.dart
