+++ lib/src/rest/rest_view.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:mdigits/common/centeredbox.dart';
-import 'package:mdigits/common/default_appbar.dart';
-import 'package:mdigits/common/default_text.dart';
-import 'package:mdigits/common/spacing_holder.dart';
-import 'package:mdigits/rest/rest_controller.dart';
+import 'package:mdigits/src/common/centeredbox.dart';
+import 'package:mdigits/src/common/default_appbar.dart';
+import 'package:mdigits/src/common/default_text.dart';
+import 'package:mdigits/src/common/spacing_holder.dart';
+import 'package:mdigits/src/rest/rest_controller.dart';
 
 class RestView extends StatelessWidget {
   RestView({super.key});
